refac: 뷰 기능 팩토리로 이전, 스코어보드 뷰 기능 -> 컨트롤러로 이전
