+++ app/src/main/java/org/tetris/scoreboard/controller/ScoreBoardController.java
@@ -2,6 +2,16 @@ package org.tetris.scoreboard.controller;
 
 import org.tetris.scoreboard.model.ScoreBoard;
 import org.tetris.scoreboard.model.ScoreInfo;
+import org.tetris.shared.BaseController;
+
+import javafx.fxml.FXML;
+import javafx.scene.control.Button;
+import javafx.scene.control.TableColumn;
+import javafx.scene.control.TableView;
+import javafx.scene.control.TextField;
+import javafx.scene.layout.AnchorPane;
+import javafx.scene.text.Text;
+
 public class ScoreBoardController extends BaseController<ScoreBoard> {
 
     private int finishScore;
@@ -16,7 +26,7 @@ public class ScoreBoardController extends BaseController<ScoreBoard> {
     }
 
     public ScoreBoard getScoreBoard(){
-        return scoreBoard;
+        return model;
     }
 
     /**
@@ -30,7 +40,38 @@ public class ScoreBoardController extends BaseController<ScoreBoard> {
 
     private void submitCurrentScore(int score, String name)
     {
-        scoreBoard.insert(new ScoreInfo(score, name));
-        scoreBoard.writeHighScoreList();
+        model.insert(new ScoreInfo(score, name));
+        model.writeHighScoreList();
+    }
+
+    @FXML private TableView<ScoreInfo> scoreTable;
+    @FXML private TableColumn<ScoreInfo, String> nameColumn;
+    @FXML private TableColumn<ScoreInfo, Integer> scoreColumn;
+
+    @FXML private AnchorPane inputPane;
+    @FXML private Text scoreText;
+    @FXML private TextField nameField;
+    @FXML private Button submitButton;
+
+
+    @Override
+    public void initialize()
+    {
+        // name, score <-> ScoreInfo
+        nameColumn.setCellValueFactory(cellData 
+            -> new javafx.beans.property.SimpleStringProperty(cellData.getValue().name()));
+        scoreColumn.setCellValueFactory(cellData 
+            -> new javafx.beans.property.SimpleIntegerProperty(cellData.getValue().score()).asObject());
+        
+        scoreText.setText(String.valueOf(getFinishScore()));
+
+        scoreTable.setItems(getScoreBoard().getScoreList());
+    }
+
+    @FXML
+    private void onSubmitPressed()
+    {
+        String playerName = nameField.getText();
+        OnSubmitClick(playerName);
     }
 }
